+++ views/amendment/create_confirm.php
@@ -73,7 +73,7 @@ echo Html::beginForm('', 'post', ['id' => 'amendmentConfirmForm']);
 echo '<div class="content">
         <div style="float: right;">
             <button type="submit" name="confirm" class="btn btn-success">
-                <span class="glyphicon glyphicon-ok-sign"></span> ' . \Yii::t('amend', 'button_submit') . '
+                <span class="glyphicon glyphicon-ok-sign"></span> ' . $amendment->getSubmitButtonLabel() . '
             </button>
         </div>
         <div style="float: left;">